requirements.txt
numpy
psycopg2_binary
pyinstaller
Shapely
schedule==1.2.2
